Update README.md
This is code is specific to allow updates WITHOUT the commonly forced replicating events function. It is common that when connecting google sheets with google calendar, that the event will update on top of the event that is being shown in google calendars. This code will allow unlimited updates to a sheet to have unlimited updates on the calendar.